Send message from extension to webpage
Add functionality to send message from extension to webpage to change
the url of the page when user wants to.
However, if the extension is already set to on during initial page load,
the url will not be changed. The page will only load if the button is
toggled for now.